ai-service/pyproject.toml
[project]
name = "mzadak-ai-service"
version = "0.1.0"
description = "MZADAK AI/ML Service (GPU)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "torch>=2.2.0",
    "transformers>=4.39.0",
    "pillow>=10.3.0",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "openai>=1.12.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
